# docs/swagger.yaml
basePath: /api
definitions:
  CreateWebhookRequest:
    properties:
      content_type:
        description: Content-Type of the response. Defaults to application/json.
        example: application/json
        type: string
      forward_url:
        description: |-
          Absolute http or https URL every captured request is also relayed to.
          Omit it, or send "", to leave forwarding off. Private and local
          addresses, such as localhost, are rejected unless the server allows
          private networks; use a public tunnel URL to reach a local server.
        example: https://example.ngrok-free.app/webhooks/stripe
        type: string
      notify_on_event:
        type: boolean
      payload:
        description: Response body
        example: '{"message":"ok"}'
        type: string
      response_code:
        description: HTTP status code the endpoint answers with, 100-599. Defaults
          to 200.
        example: 200
        type: integer
      response_delay:
        description: Milliseconds to wait before answering, 0-30000
        example: 0
        type: integer
      response_headers:
        additionalProperties:
          type: string
        description: Extra response headers
        type: object
      title:
        description: Title of the webhook
        example: Payment events
        type: string
    required:
    - title
    type: object
  ErrorResponse:
    properties:
      error:
        example: webhook not found
        type: string
    type: object
  UpdateWebhookRequest:
    properties:
      content_type:
        example: application/json
        type: string
      forward_url:
        description: |-
          Absolute http or https URL every captured request is also relayed to.
          null or "" turns forwarding off. Private and local addresses, such as
          localhost, are rejected unless the server allows private networks; use
          a public tunnel URL to reach a local server.
        example: https://example.ngrok-free.app/webhooks/stripe
        type: string
        x-nullable: true
      notify_on_event:
        type: boolean
      payload:
        example: '{"message":"ok"}'
        type: string
      response_code:
        example: 200
        type: integer
      response_delay:
        example: 0
        type: integer
      response_headers:
        additionalProperties:
          type: string
        description: Replaces all response headers
        type: object
      title:
        example: Payment events
        type: string
    type: object
  Webhook:
    properties:
      content_type:
        type: string
      created_at:
        type: string
      forward_url:
        description: Where captured requests are relayed; null when forwarding is
          off
        example: https://example.ngrok-free.app/webhooks/stripe
        type: string
        x-nullable: true
      id:
        type: string
      notify_on_event:
        type: boolean
      payload:
        type: string
      requests:
        items:
          $ref: '#/definitions/WebhookRequest'
        type: array
      response_code:
        type: integer
      response_delay:
        description: milliseconds
        type: integer
      response_headers:
        additionalProperties:
          type: string
        type: object
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  WebhookRequest:
    properties:
      body:
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      method:
        type: string
      path:
        example: /orders/42
        type: string
      query:
        additionalProperties:
          type: string
        type: object
      received_at:
        type: string
      webhook_id:
        type: string
    type: object
info:
  contact:
    email: william@srninety.one
    name: William Muli
  description: Manage your Webhook Tester endpoints programmatically. Authenticate
    every request with the X-API-Key header; your key is under API access in your
    [workspace](/).
  title: Webhook Tester API
  version: "1.0"
paths:
  /webhooks:
    get:
      description: List webhooks and their most recent requests
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Webhook'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List webhooks
      tags:
      - Webhooks
    post:
      consumes:
      - application/json
      description: Returns the details of the created webhook
      parameters:
      - description: Webhook body
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/CreateWebhookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a webhook
      tags:
      - Webhooks
  /webhooks/{id}:
    delete:
      description: Deletes a webhook and its captured requests
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a webhook
      tags:
      - Webhooks
    get:
      description: Get a webhook by ID along with its requests
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Webhook'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get webhook by ID
      tags:
      - Webhooks
    patch:
      consumes:
      - application/json
      description: Changes only the fields included in the body. PUT is accepted as
        an alias of PATCH.
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a webhook
      tags:
      - Webhooks
    put:
      consumes:
      - application/json
      description: Changes only the fields included in the body. PUT is accepted as
        an alias of PATCH.
      parameters:
      - description: Webhook ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: webhook
        required: true
        schema:
          $ref: '#/definitions/UpdateWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Webhook'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a webhook
      tags:
      - Webhooks
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"